Initial commit
review-strategy plugin, A plugin to control various ways to review
changes. Although I've only implemented one strategy, the purpose
of this plugin is to allow developers to implement many different
review strategies.

Done:
 * Adds a per-user preference to select the review strategy
 * BLIND strategy hides the following information on the change view:
   author, commiter, reviewer(s), change history summaries.

Note - 'expand all' button will show reviewer's votes and comments.
I've left this as-is on purpose so that users can get a *sneek peek*
at the votes.